Add serial console

+++ include/sprintf.h
@@ -6,16 +6,9 @@
 #include <stdarg.h>
 #include <stddef.h>
 
-/* formatted output to a fixed-length string */
-size_t snprintf(char *restrict s, size_t n, const char *restrict fmt, ...) __format(printf, 3, 4) __weak;
-
-/* formatted output to a string */
-size_t sprintf(char *restrict s, const char *restrict fmt, ...) __format(printf, 2, 3) __weak;
-
-/* formatted output to a fixed-length string */
-size_t vsnprintf(char *restrict s, size_t n, const char *restrict fmt, va_list va) __weak;
-
-/* formatted output to a string */
-size_t vsprintf(char *restrict s, const char *restrict fmt, va_list va) __weak;
+int snprintf(char *restrict s, size_t n, const char *restrict fmt, ...) __format(printf, 3, 4) __weak;
+int sprintf(char *restrict s, const char *restrict fmt, ...) __format(printf, 2, 3) __weak;
+int vsnprintf(char *restrict s, size_t n, const char *restrict fmt, va_list va) __weak;
+int vsprintf(char *restrict s, const char *restrict fmt, va_list va) __weak;
 
 #endif /* __INCLUDE_SPRINTF_H__ */
